backend/requirements.txt
# Datasheet Part Selector Backend

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
docling>=2.0.0
google-generativeai>=0.3.0
sse-starlette>=1.8.0
aiosqlite>=0.19.0
httpx>=0.24.0

